Se crea las notificaciones del archivo plano de claims y se documenta el ultimo controlador
